Moved new components into folders and updated references
