+++ pages/assets.tsx
@@ -1,6 +1,7 @@
 import type { NextPage } from "next";
 import Layout from "../components/Layout";
-import Pagination from "../components/Pagination";
+import Pagination from "../components/Pagination/Pagination";
+import SummaryPanel from "../components/SummaryPanel/SummaryPanel";
 
 const Assets: NextPage = () => {
   return (
@@ -11,16 +12,7 @@ const Assets: NextPage = () => {
           <div><p className="lbh-body-m">Showing 8 of 18 vacancies</p></div>
           <ul className="lbh-list">
             <li>
-              <div className="lbh-summary-panel">
-                <div className="lbh-summary-panel__content">
-                  <a href="#" className="lbh-link">1 bedroom, 3 bathrooms</a>
-                  <p className="lbh-body-m">Shared bathroom, shared kitchen</p>
-                  <p className="lbh-body-m govuk-!-margin-bottom-3">Unknown access, unknown floor</p>
-                  <p className="lbh-body-s">Hackney Hostel - Hostel LSE, £32 a night</p>
-                  <p className="lbh-body-s">Room 1, 11 Meynell Road, Hackney. E9 7AP.</p>
-                </div>
-                <img className="lbh-summary-panel__img" src="https://hackneywebsite-wp-mediafiles.s3.eu-west-2.amazonaws.com/wp-content/uploads/1920_1920-hackneytownhallwithlights-2.jpg" alt="Hackney at Christmas"></img>
-              </div>
+              <SummaryPanel />
             </li>
           </ul>
           <Pagination />
